refactor: 키워드 관련 파일 폴더로 이동
